+++ com.rustin.rs3.leveltracker2.0.sdPlugin/ui/rs3-gim-rank.js
@@ -8,9 +8,9 @@
     showRank: true,
     showLevel: false,
     showXp: false,
-    titleBold: false,
+    titleBold: true,
     titleColor: "#FFFFFF",
-    titleSize: 22,
+    titleSize: 24,
     linkToMain: false
   };